Moved inplace_change function to plugin.py
