+++ yin_yang/plugins/atom.py
@@ -1,38 +1,17 @@
 import os
 import pwd
 import re
-from yin_yang.plugins.plugin import Plugin
+from yin_yang.plugins.plugin import Plugin, inplace_change
 
 # aliases for path to use later on
 user = pwd.getpwuid(os.getuid())[0]
-path = "/home/"+user+"/.atom/config.cson"
-
-
-def inplace_change(filename, old_string, new_string):
-    """@params: config - config to be written into file
-                path - the path where the config is will be written into.
-                    Defaults to the default path
-    """
-    # Safely read the input filename using 'with'
-    with open(filename) as f:
-        s = f.read()
-        if old_string not in s:
-            print('"{old_string}" not found in {filename}.'.format(**locals()))
-            return
-
-    # Safely write the changed content, if found in the file
-    with open(filename, 'w') as f:
-        print(
-            'Changing "{old_string}" to "{new_string}" in {filename}'
-            .format(**locals()))
-        s = s.replace(old_string, new_string)
-        f.write(s)
+path = "/home/" + user + "/.atom/config.cson"
 
 
 def get_old_theme(settings):
     """returns the theme which is currently used
        uses regex to find the currently used theme
-       i excpect that themes follow this pattern
+       i expect that themes follow this pattern
        XXXX-XXXX-ui     XXXX-XXXX-syntax
     """
     with open(settings, "r") as file:
